+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `fail` creates a labelled failure, which is like a throw for parsers
 - `failLabel(parser, label)` is a shortcut for `or(parser, fail(label))`
 - `cstToIst` transforms a CST to an IST
+- `istTransformer` uses `rule`s to transform your IST into a real AST
 
 [unreleased]: https://github.com/unleashy/vahv/compare/v0.1.0...HEAD